requirements.txt
pytest
psycopg2-binary
sqlalchemy
python-dotenv
flask
pydantic